chore(miners): inline MinerPickerModal back into ManageMinersModal
The shared wrapper had a single transitive caller and the planned
/fleet "Add miners → site/building" reuse case didn't ship in this PR.
Same YAGNI pattern as the RackPickerModal cleanup. ManageMinersModal
returns to its pre-extraction shape (slot cap + disable-rows predicate
+ "Continue" button).

If/when the /fleet Add miners flow lands, re-extract a shared picker
then — likely as part of unifying SearchMinersModal + the multi-select
flow into one component with a `selectionMode` prop.
