--- client/src/protoFleet/components/MinerPickerModal/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./MinerPickerModal";
-export type { MinerPickerConfirmPayload } from "./MinerPickerModal";
